cadastro de servicos

+++ src/main/java/br/com/alura/barbeariaonline/model/Servico.java
@@ -42,7 +42,7 @@ public class Servico {
         return valor;
     }
 
-    public void setValor(float valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
